+++ src/DeepWoods/Stuff/DeepWoodsMineCart.cs
@@ -1,10 +1,9 @@
 ﻿
 using Microsoft.Xna.Framework;
+using StardewModdingAPI;
 using StardewValley;
-using StardewValley.Menus;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
+using xTile.Dimensions;
+using xTile.Layers;
 using xTile.Tiles;
 using static DeepWoodsMod.DeepWoodsSettings;
 
@@ -13,6 +12,8 @@ namespace DeepWoodsMod.Stuff
     public class DeepWoodsMineCart
     {
         public static Vector2 MineCartLocation => new Vector2(Settings.Map.RootLevelEnterLocation.X - Settings.Map.ExitRadius - 2 - 6, Settings.Map.RootLevelEnterLocation.Y + 5);
+
+        public static string DeepWoodsMineCartId => "maxvollmer.deepwoods.DeepWoodsMineCart";
 
         public static void AddDeepWoodsMineCart(DeepWoods location)
         {
@@ -29,27 +30,6 @@ namespace DeepWoodsMod.Stuff
                 AddTrackTile(location, MineCartLocation, 0, -i, 1076);  // track
             }
             AddTrackTile(location, MineCartLocation, 0, 1, 1075);       // end of track
-
-            // If MinecartPatcher is installed, ensure this minecart is loaded
-            // (we then inject it in DeepWoodsContentAdder)
-            if (ModEntry.GetHelper().ModRegistry.IsLoaded("mod.kitchen.minecartpatcher"))
-            {
-                var mod = ModEntry.GetHelper().ModRegistry.Get("mod.kitchen.minecartpatcher");
-                if (mod != null)
-                {
-                    // mod is actually of type ModMetadata
-                    var modField = mod.GetType()?.GetField("Mod", BindingFlags.Instance | BindingFlags.Public);
-                    if (modField != null)
-                    {
-                        var modEntry = modField.GetValue(mod);
-                        if (modEntry != null)
-                        {
-                            var loadDataMethod = modEntry.GetType()?.GetMethod("LoadData", BindingFlags.Instance | BindingFlags.Public);
-                            loadDataMethod?.Invoke(modEntry, null);
-                        }
-                    }
-                }
-            }
         }
 
         private static Tile GetOrCreateTile(DeepWoods location, string layerName, int tileX, int tileY, int tileIndex)
@@ -78,124 +58,8 @@ namespace DeepWoodsMod.Stuff
             int tileY = (int)(tileLocation.Y + y);
             var tile = GetOrCreateTile(location, "Buildings", tileX, tileY, tileIndex);
             tile.Properties.Remove("Action");
-            tile.Properties.Add("Action", "MinecartTransport");
+            tile.Properties.Add("Action", $"MinecartTransport Default {DeepWoodsMineCartId}");
             tile.TileIndex = tileIndex;
-        }
-
-        public static void InjectDeepWoodsMineCartIntoGame()
-        {
-            if (Game1.MasterPlayer.mailReceived.Contains("ccBoilerRoom")
-                && DeepWoodsState.LowestLevelReached >= 1
-                && Game1.dialogueUp
-                && Game1.currentLocation.lastQuestionKey == "Minecart"
-                && Game1.activeClickableMenu is DialogueBox dialogBox
-                && dialogBox is not PatchedDialogueBox
-                && IsMineCartOptions(dialogBox.responses))
-            {
-                if (Game1.currentLocation is DeepWoods)
-                {
-                    Game1.activeClickableMenu = new PatchedDialogueBox(dialogBox, "Mines", Game1.content.LoadString("Strings\\Locations:MineCart_Destination_Mines"));
-                }
-                else
-                {
-                    Game1.activeClickableMenu = new PatchedDialogueBox(dialogBox, "DeepWoods", I18N.DeepWoodsMineCartText);
-                }
-            }
-        }
-
-        private static bool IsMineCartOptions(Response[] responses)
-        {
-            if (responses.Length < 3)
-            {
-                return false;
-            }
-
-            bool canHasQuarry = Game1.MasterPlayer.mailReceived.Contains("ccCraftsRoom");
-
-            bool hasMines = responses.Where(r => r.responseKey == "Mines").Any();
-            bool hasTown = responses.Where(r => r.responseKey == "Town").Any();
-            bool hasBus = responses.Where(r => r.responseKey == "Bus").Any();
-
-            if (canHasQuarry)
-            {
-                bool hasQuarry = responses.Where(r => r.responseKey == "Quarry").Any();
-                int count = CountTrue(hasMines, hasTown, hasBus, hasQuarry);
-                if (count < 3)
-                {
-                    return false;
-                }
-            }
-            else
-            {
-                int count = CountTrue(hasMines, hasTown, hasBus);
-                if (count < 2)
-                {
-                    return false;
-                }
-            }
-
-            return true;
-        }
-
-        private static int CountTrue(params bool[] args)
-        {
-            return args.Count(t => t);
-        }
-
-        private class PatchedDialogueBox : DialogueBox
-        {
-            public PatchedDialogueBox(DialogueBox original, string newOptionKey, string newOptionText)
-                : base(original.dialogues[0], InsertResponse(original.responses.ToList(), new Response(newOptionKey, newOptionText)))
-            {
-            }
-
-            private static Response[] InsertResponse(List<Response> responses, Response newResponse)
-            {
-                responses.Insert(responses.Count - 1, newResponse);
-                return responses.ToArray();
-            }
-
-            public override void receiveLeftClick(int x, int y, bool playSound = true)
-            {
-                if (transitioning)
-                {
-                    return;
-                }
-
-                if (safetyTimer > 0)
-                {
-                    return;
-                }
-
-                if (selectedResponse < 0 ||selectedResponse >= responses.Length)
-                {
-                    return;
-                }
-
-                if (responses[selectedResponse].responseKey != "DeepWoods")
-                {
-                    base.receiveLeftClick(x, y, playSound);
-                    return;
-                }
-
-                questionFinishPauseTimer = (Game1.eventUp ? 600 : 200);
-                transitioning = true;
-                transitionInitialized = false;
-                transitioningBigger = true;
-                Game1.dialogueUp = false;
-                Game1.player.Halt();
-                Game1.player.freezePause = 700;
-                Game1.warpFarmer("DeepWoods", 105, 80, 1);
-                Game1.warpFarmer("DeepWoods", (int)MineCartLocation.X, ((int)MineCartLocation.Y) + 1, 2);
-                DeepWoodsManager.currentWarpRequestName = "DeepWoods";
-                DeepWoodsManager.currentWarpRequestLocation = new Vector2(MineCartLocation.X * 64, (MineCartLocation.Y + 1) * 64);
-                Game1.playSound("smallSelect");
-                selectedResponse = -1;
-                if (Game1.activeClickableMenu != null && Game1.activeClickableMenu.Equals(this))
-                {
-                    beginOutro();
-                }
-            }
         }
     }
 }